chores: - only let authors delet and update their post - use current user decorator to populate user data

+++ blog-api/src/posts/posts.controller.ts
@@ -7,20 +7,24 @@ import {
   Param,
   Delete,
   Query,
+  UseGuards,
 } from '@nestjs/common';
 import { PostsService } from './posts.service';
 import { CreatePostDto } from './dto/create-post.dto';
 import { UpdatePostDto } from './dto/update-post.dto';
 import { QueryPostDto } from './dto/filter-post.dto';
 import { Types } from 'mongoose';
+import { JwtAuthGuard } from '../auth/jwt-auth.guard';
+import { CurrentUser } from '../common/decorators/user.decorator';
 
 @Controller('posts')
+@UseGuards(JwtAuthGuard)
 export class PostsController {
   constructor(private readonly postsService: PostsService) {}
 
   @Post()
-  create(@Body() createPostDto: CreatePostDto) {
-    return this.postsService.create(createPostDto);
+  create(@CurrentUser() user, @Body() createPostDto: CreatePostDto) {
+    return this.postsService.create({ ...createPostDto, author: user._id });
   }
 
   @Get()
@@ -35,14 +39,15 @@ export class PostsController {
 
   @Patch(':id')
   update(
+    @CurrentUser() user,
     @Param('id') _id: Types.ObjectId,
     @Body() updatePostDto: UpdatePostDto,
   ) {
-    return this.postsService.update(_id, updatePostDto);
+    return this.postsService.update({ _id, updatePostDto, author: user._id });
   }
 
   @Delete(':id')
-  remove(@Param('id') _id: Types.ObjectId) {
-    return this.postsService.remove({ _id });
+  remove(@CurrentUser() user, @Param('id') _id: Types.ObjectId) {
+    return this.postsService.remove({ _id, author: user._id });
   }
 }
